# docs/swagger.yaml
basePath: /
definitions:
...
  schedule.Faculty:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  schedule.Group:
    properties:
      course:
        type: integer
      degree:
        description: бакалавриат / магистратура
        type: string
      facultyID:
        type: integer
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      specialization:
        type: string
      yearStart:
        type: integer
    type: object
  schedule.Lesson:
    properties:
      classroom:
        type: string
      createdAt:
        type: string
      dayOfWeek:
        type: string
      endTime:
        type: string
      groupName:
        type: string
      id:
        type: integer
      isEvenWeek:
        type: boolean
      isOnline:
        type: boolean
      startTime:
        type: string
      subject:
        type: string
      teacher:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Backend API for the TechUp university application.
  title: TechUp API
  version: "1.0"
paths:
...
  /api/v1/admin/faculty:
    post:
      consumes:
      - application/json
      description: Adds a new faculty
      parameters:
      - description: Faculty data
        in: body
        name: faculty
        required: true
        schema:
          $ref: '#/definitions/schedule.Faculty'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.Faculty'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add faculty
      tags:
      - admin-schedule
  /api/v1/admin/faculty/{id}:
    delete:
      description: Deletes faculty by ID
      parameters:
      - description: Faculty ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete faculty
      tags:
      - admin-schedule
    put:
      consumes:
      - application/json
      description: Updates faculty by ID
      parameters:
      - description: Faculty ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated faculty
        in: body
        name: faculty
        required: true
        schema:
          $ref: '#/definitions/schedule.Faculty'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.Faculty'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update faculty
      tags:
      - admin-schedule
  /api/v1/admin/group:
    post:
      consumes:
      - application/json
      description: Adds a new group
      parameters:
      - description: Group data
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/schedule.Group'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.Group'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add group
      tags:
      - admin-schedule
  /api/v1/admin/group/{id}:
    delete:
      description: Deletes group by ID
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete group
      tags:
      - admin-schedule
    put:
      consumes:
      - application/json
      description: Updates an existing group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated group
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/schedule.Group'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.Group'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update group
      tags:
      - admin-schedule
  /api/v1/admin/lesson:
    post:
      consumes:
      - application/json
      description: Creates a new lesson in the schedule
      parameters:
      - description: Lesson data
        in: body
        name: lesson
        required: true
        schema:
          $ref: '#/definitions/schedule.Lesson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.Lesson'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a new lesson
      tags:
      - admin-schedule
  /api/v1/admin/lesson/{id}:
    delete:
      description: Deletes a lesson by ID
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete lesson
      tags:
      - admin-schedule
    put:
      consumes:
      - application/json
      description: Updates an existing lesson
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated lesson data
        in: body
        name: lesson
        required: true
        schema:
          $ref: '#/definitions/schedule.Lesson'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schedule.Lesson'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update lesson by ID
      tags:
      - admin-schedule
  /api/v1/schedule/faculties:
    get:
      description: Returns all faculties
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schedule.Faculty'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all faculties
      tags:
      - schedule
  /api/v1/schedule/groups:
    get:
      description: Returns list of groups
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schedule.Group'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all groups
      tags:
      - schedule
  /api/v1/schedule/lessons:
    get:
      description: Returns a list of all lessons
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schedule.Lesson'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all lessons
      tags:
      - schedule
  /api/v1/schedule/search:
    get:
      description: Search lessons by group, teacher, classroom, day, time range, or
        even/odd week. At least one query parameter is required.
      parameters:
      - description: Group name
        in: query
        name: group
        type: string
      - description: Teacher name
        in: query
        name: teacher
        type: string
      - description: Classroom
        in: query
        name: classroom
        type: string
      - description: Day of the week
        in: query
        name: day
        type: string
      - description: Start time in HH:MM format
        in: query
        name: from
        type: string
      - description: End time in HH:MM format
        in: query
        name: to
        type: string
      - description: Even week (true/false)
        in: query
        name: is_even_week
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/schedule.Lesson'
            type: array
        "400":
          description: At least one query parameter is required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Search lessons schedule
      tags:
      - schedule
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"